Task 16. Fibonacci 98

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 	int i;
-	double n1, n2, sum;
+	unsigned double n1, n2, sum;
 
 	n1 = 1;
 	n2 = 2;
